Move atomically to unstable::sync, and document what it actually does. Close #7872.
